Addressed VS 2008 warning again.
git-svn-id: svn://stormbringer.nateko.lu.se/svn/LPJ-GUESS/branches/fire_blaze_merge@7948 70805675-57b6-444b-bb00-18ecffcb4be8

+++ modules/simfire.cpp
@@ -178,7 +178,7 @@ void simfire_biome_mapping(Gridcell& gridcell) {
 	int count_max=0; // maximum of 'count'
 	// find and save most common biome number
 	for (biome=0; biome<NFIREBIOMES; biome++) count[biome]=0;
-	for (int idx = 0; idx < biomes.size(); idx++) {
+	for (int idx = 0; idx<(int)biomes.size(); idx++) {
 		count[biomes[idx]]++;
 	}
 	for (biome=0; biome<NFIREBIOMES; biome++) {
